Change the PORT environment variable for keter.
I _believe_ keter expects the webapp to accept the `PORT` environment
variable in order to set the port the webapp is listening on.

Kucipong originally used the `KUCIPONG_PORT` environment variable instead
of the `PORT` environment variable.  This PR changes it so that Kucipong
uses the `PORT` environment variable so it works well with keter.

@arowM do you mind if I try releasing this onto the staging server?  I
haven't tested it locally, so I am not 100% sure it works.

Closes #94.