Add I18n instance for CouponType.
This adds an `I18n` instance for `CouponType`.

It also changes some of the pug files to use `label` to show the
`couponType` value.

This changes the way the `CouponType`s are stored in the database.  In
order to get the server to load with this commit, you must delete all
existing coupons:

```sh
$ psql -U kucipong -d kucipong -h 127.0.0.1
kucipong=> delete from coupon;
```

Fixes #83.